:recycle: Refactor: 도메인(상품, 주문) 패키지 구조 수정
