+++ src/main/java/com/umc/thegoods/domain/images/ReviewImg.java
@@ -1,7 +1,7 @@
 package com.umc.thegoods.domain.images;
 
-import com.umc.thegoods.domain.Review;
 import com.umc.thegoods.domain.common.BaseDateTimeEntity;
+import com.umc.thegoods.domain.item.Review;
 import lombok.*;
 
 import javax.persistence.*;
