+++ src/main/java/com/umc/thegoods/domain/item/ItemOption.java
@@ -1,6 +1,7 @@
-package com.umc.thegoods.domain;
+package com.umc.thegoods.domain.item;
 
 import com.umc.thegoods.domain.common.BaseDateTimeEntity;
+import com.umc.thegoods.domain.order.OrderDetail;
 import lombok.*;
 
 import javax.persistence.*;
