+++ src/main/java/com/umc/thegoods/domain/order/Cart.java
@@ -1,7 +1,7 @@
-package com.umc.thegoods.domain;
+package com.umc.thegoods.domain.order;
 
 import com.umc.thegoods.domain.common.BaseDateTimeEntity;
-import com.umc.thegoods.domain.enums.ReviewStatus;
+import com.umc.thegoods.domain.item.ItemOption;
 import com.umc.thegoods.domain.user.User;
 import lombok.*;
 
@@ -12,31 +12,20 @@ import javax.persistence.*;
 @Builder
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @AllArgsConstructor
-public class Review extends BaseDateTimeEntity {
+public class Cart extends BaseDateTimeEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
-    @Column(nullable = false, length = 1500)
-    private String text;
-
-    @Column(nullable = false, length = 1)
-    private Integer score;
-
-    @Enumerated(EnumType.STRING)
-    @Column(columnDefinition = "VARCHAR(10)", nullable = false)
-    private ReviewStatus status;
+    @Column(nullable = false)
+    private Integer amount;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "USER_ID", nullable = false)
     private User user;
 
     @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "ITEM_ID", nullable = false)
-    private Item item;
-
-    @OneToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "OPTION_ID", nullable = false)
     private ItemOption itemOption;
 }
