+++ src/main/java/com/umc/thegoods/mapping/ViewSearch/ItemView.java
@@ -1,7 +1,7 @@
 package com.umc.thegoods.mapping.ViewSearch;
 
-import com.umc.thegoods.domain.Item;
 import com.umc.thegoods.domain.common.BaseDateTimeEntity;
+import com.umc.thegoods.domain.item.Item;
 import com.umc.thegoods.domain.user.User;
 import lombok.*;
